Test topics added
