+++ tests/config/databaseTest.js
@@ -29,4 +29,6 @@ const dropCollections = async () => {
     }
 };
 
+const createDummyUsers = async () => {};
+
 module.exports = { connectDB, dropDB, dropCollections };
